feat: deep simulation stack — variable genome, metabolic networks, proto-proteins, codons, multicellularity, cancer therapy
VG-1→6: VariableGenome (4-32 genes), Kleiber maintenance cost, gene duplication/deletion,
         epigenetic gating, expression mapping, serialization (62 tests)

MGN-1→7: gene→ExergyNode, DAG topology inference, MetabolicGraph from genome,
          node competition (Axiom 3), Hebbian rewiring, internal catalysis (80 tests)

PF-1→5: HP lattice protein fold, contact map, active sites, catalytic function inference,
         compute_protein_phenotype cache (27 tests)

PD-1→5: CodonGenome (tripletes), CodonTable (64→8 amino, evolucionable), translation pipeline,
         silent mutations, neutral drift, batch wiring (28 tests)

MC-1→5: cell adhesion (freq×distance), colony detection (Union-Find), positional signaling,
         differential expression (borde=defensa, interior=metabolismo), batch wiring (33 tests)

Cancer Therapy: Hill pharmacology, frequency-selective drug, quiescent stem cells,
                normal regeneration, microenvironment, immune cells, PK ramp,
                CSV export, calibration-ready (19 tests)

Infrastructure:
- MAX_ENTITIES 64→128, alive_mask u64→u128
- 13 use case binaries (fermi, speciation, cambrian, versus, museum, fossil, petri, music, mesh, etc.)
- Layer fixes: MobaIdentity Vec→bitfield, Grimoire Vec→array, AlchemicalForge Vec→array
- SV-1: apply_input() wiring (InputCommand → WillActuator)
- Centralized gaussian_frequency_alignment + sanitize_unit in determinism.rs
- Observability: gene_count, metabolic_graph_rate, protein_rate, codon_count, multicellular_rate
- Updated: README, CHANGELOG, CLAUDE.md, all docs

2940 tests, 0 failed. 87K+ LOC. 64 sprints completed. 8 axioms verified. Zero hardcode.

+++ src/blueprint/equations/batch_stepping.rs
@@ -56,12 +56,12 @@ pub fn locomotion_n_ticks(qe: f32, speed: f32, terrain_factor: f32, n: u32) -> f
 /// If isolated, analytical stepping is exact (no interactions to miss).
 pub fn is_isolated(
     positions: &[[f32; 2]],
-    alive_mask: u64,
+    alive_mask: u128,
     entity_idx: usize,
     range_sq: f32,
 ) -> bool {
     let pos = positions[entity_idx];
-    let mut mask = alive_mask & !(1u64 << entity_idx);
+    let mut mask = alive_mask & !(1u128 << entity_idx);
     while mask != 0 {
         let j = mask.trailing_zeros() as usize;
         mask &= mask - 1;
